Copied RRB Vector code from FSharpx.Collections
Copied my implementation of RRB Vector from FSharpx.Collections, so that
I can release it for .Net Standard 1.6 and 2.0 without having to wait
for the FSharpx.Collections release process (which is much slower due to
having so much more code to test).

This will eventually get merged back into FSharpx.Collections, of course.

Tests will come in the next commit

Add Expecto tests for RRB Vector

Fix type mismatch error on empty root nodes

Apparently when emptyNode is referenced, it's getting a node of the
wrong type somehow. TODO: Check if calling it with emptyNode<'T> would
work properly (remove this commit and try that instead).