Try to fix the singleton-node problem
When a node has just one child, it's automatically considered a full
node... which causes problems when AppendChild looks at it and says
"Hey, this is a full node so it'll stay a full node after appending" and
you get a Node([|RRBNode; RRBNode|]), which is wrong.

This commit is one solution to the problem: a special class for nodes
with one child, which will convert itself to a full node only if its
child is also full (not an RRBNode). We could do more with this
SingletonNode class, like optimize a few operations, but for now this is
enough to make the failing test pass, while still being O(1) on pushes.