Adjust vector property checks to let test pass
After further thinking about the required invariant, I realized that
this test should really be passing. If the root node of a tree is a
4-array of (full, full, full, relaxed) then the root can safely be
allowed to be a full node, as radix indexing will always choose the
right path. The actual invariant I should be checking is that full
nodes, if they have any non-full children, MUST have those children as
the last node.

With the property check adjusted to match that actual invariant, this
test now passes.