WIP on transients, nowhere close to finished
Basic idea is we'll have "expanded" versions of basic nodes and RRB
nodes, whose array will be always 32 in size. In any transient vector,
only the right spine will be expanded, so that it can be converted to a
persistent (which trims the arrays down to size) in effective O(1) time.