WIP: More progress on transients
Now we have TakeChildren and SkipChildren methods on nodes which can be
used in the Take and Skip implementations. Note that the implementation
of Split() will have to be carefully considered. Currently the first
Take() would mutate the tree so that Skip() would fail. We'll have to
rewrite it so that Skip() is called first (making a brand-new transient
tree that shares no roots with the current one) and then Take could be
called second, mutating the current tree. We'll need lots of unit tests
around that scenario to make sure we don't end up sharing any nodes
between the two related trees after the split, otherwise things could
break once you start mutating either tree.