VERY WIP: Generic nodes need two type parameters
Many type errors remaining as this is not at all finished, but the basic
idea seems workable. Generic nodes need two type parameters, one for the
type of the child items (which we will call 'C) and one for the type of
the content items (which we have been calling 'T). In a leaf node, the
types 'T and 'C are the same type. In a twig node, 'C is Node<'T,'T>. In
a branch node, 'C is Node<Node<'T, 'T>, 'T>, and so on up the tree
levels. In principle, this should work. In practice, well, we'll see.