WIP: Better implementation of nodes
This time I'm starting from scratch so I don't get lost.

I think the best approach is going to be to implement what is
essentially a the FullNode/TreeNode/LeafNode DU from my first attempt,
but as a classic OOP-style class hierarchy. That will cause a bit more
complexity in some parts, but will simplify other parts (for example, no
more distinction between the LeafNode DU case and the LeafNode record,
which was causing confusion). This will also allow me to implement
transience, which the DU-and-record case would have made tricky (since
transience needs mutability).

Inheritance diagram: "Node", which will eventually be renamed to
"RRBNode" once that can no longer cause confusion, is the abstract base
class. "RRBFullNode" is the class previously known as Node or FullNode,
"RRBRelaxedNode" is the class previously known as RRBNode or TreeNode,
and "RRBLeafNode" is the class previously known as LeafNode (or just a
bare array in my attempt that didn't quite work). All three of these
derive from Node (or perhaps RRBRelaxedNode derives from RRBFullNode).
The leaf class has an array of 'T items, while the other node classes
have an array of Node<'T> children. No boxing needed with this approach!