WIP on insertion into better leaf nodes
I'm discovering needed node methods as I write; for example, there's a
need for a function that will split nodes into two arrays, which I'll
make in my next commit. This one preserves the state of functions like
RemoveLeftmostChildren, which I'm going to tear apart and rewrite in the
next commit.