Add basic implementation of rebalance helper funcs
There are two functions here, but the second one is the one I'll end up
keeping. It does a single O(N) sweep through an array to find the
smallest subarray with sum of at least N. Make an array of "gaps" in
child nodes ("gap" size is block size minus node count), and then
setting N = block size means we're finding the least amount of work
needed to reduce by one slot after the rebalance. If we then set N to be
two times block size, that would be the least amount of work to reduce
by two slots after the rebalance, which could potentially add up to some
time savings later on. We won't do that right away, but we'll experiment
with it once this is working.