Simplify SetNodeSize and test for its correctness
We no longer use SetNodeSize as a way to "zero out" extra data in shrinking
nodes; instead we make it the job of the low-level node-manipulation
functions to zero out the appropriate indices. We also have a test to check
that all expanded nodes have exactly the right amount of data in them (and
in their size tables, but it's the Children array that matters more since
that allows for garbage collection).