Revert "Stop using sequenceEqual in vector property tests"
This reverts commit 398dfffe793505c0e6da8433ca5436bb939c0ded.

We will eventually want to use this commit, but let's wait until we've
switched RRBVector over to using the new nodes before we make this change.
That way we'll have more confidence that the test suite passes all the
same tests it used to pass, and we'll also have better data for how long
it takes to run the tests before and after the change.