Make ConcatNodes do less work most of the time
Only if rebalancing is needed will we rewrite a node; otherwise we'll
just return the two nodes unchanged. Still to be determined is what
we'll do if the two nodes can easily be combined into a single node
(i.e. their combined size is less than Literals.blockSize). On the one
hand, combining them into a single node will create more work, and
rebalancing at the one-level-up location would probably be enough.
On the other hand, the one-level-up location will only be rebalanced
once it's at least 4 long, because 3 long is at most 2 more than the
minimum, and that's quite a bit of time where the higher-up node is
less than perfect.