Remove one level of indirection in vector length
In internal code, we always use this.Count rather than this.Length
so that we have one less function call. Minor optimization, but worth
it since it adds no complexity. (Rather, it removes some complexity
since we don't have to remember that only this.Count is mutable in
transients -- this.Length is never mutable since it's a property
exposed to users).