WIP Partway through code audit
Audited up to but not including this.Append. Also decided that there
really is a use for GetEditableNodeOfLengthN, but probably only for
leaf nodes. Once I finish the audit and write tests, go and tweak that
API so it returns an editable array, not just an editable node. That
way I won't be creating a node just to copy a tail array.