Tweak new transient-vector test, reducing size
Problem shows up as soon as we get to size (M*M*M*2)+M+1, which in
the case of M=32 is equal to 65569. From 1 to 65568 the vector passes
all property checks, but once you push one more item then there's an
expanded node in the tree in a position where it shouldn't be.