Fix bug, so regression test passes now
There were two separate bugs that both triggered the same test failure.
Both are fixed now. One was the fact that MkNodeForRebalance, if it had
to create a new node, wasn't creating an expanded node when it should.
The other was that it really was newRight, not newLeft, that needed to
be the basis for NewParent inside the append algorithm.