Remove unused code
Turns out GetEditableNodeOfBlockSizeLength is only useful on leaf nodes
which have been promoted to become the new tail of a transient vector,
such as after a split or popping the last item of a 1-item tail.

Next commit will change the function name and make it return an array
rather than a node, so we don't have to create a new node instance that
we're just immediately going to throw away since we're only interested
in its Items.