Fix bug in InsertAndSlide(Left/Right), add tests
The special case in InsertAndSlideChildrenLeft needed to be in Right
instead. There's no way for that special case to be hit in the Left
version, but in the Right version we had a couple corner cases where the
current node was full and the right sibling was almost full (length of
blockSize - 1). Now there are regression tests covering both those
formerly-buggy corner cases.