Fix test failures on mapFoldBack
It turns out that mapFoldBack failures were caused by RRBVector.rev
returning the wrong result, but that wasn't the actual bug. The actual
bug was in GetItem on transients, where we were using `tail` (the
constructor parameter) instead of `this.Tail` (the instance property).
If the vector's tail is no longer the same array that was passed to the
constructor, which can happen when you push into a full tail and it gets
turned into a leaf, then GetItem on an index past TailOffset was
fetching the wrong value because it was fetching from the wrong array.