Improve mapFold and mapFoldBack
Need to check other functions as well for places where we might be
constraining the result types unnecessarily. In this case, the
optimization was constraining: we tried to take a shortcut for empty
vectors, but that constrained the result type to be the same as the
input type (because we were returning the input vector unchanged if it
was empty), when in `mapFold` there's no need for the output type to
have anything at all to do with the input type. Other functions (map,
scan, etc.) need to be checked for similar optimization-caused
constraints so we have fully-expressive functions in the API.