Simplify string representation of nodes
There's really no need for a long StringRepr any more, and it's
interfering with debugging: calculating the StringRepr is often making
the VS Code debugger time out and fail when it tries to expand a node
more than two levels tall. So now the StringRepr will be short no matter
the contents of the node.