Remove unused code
Now that we shrink the right spine of a left node *before* appending the
nodes in rebalance operations, there's no need to shrink all the
children afterwards. (It was previously needed because we didn't know
where the left node would end up in the rebalanced tree, but it didn't
work properly anyway: there were still some scenarios where a grandchild
of the node wouldn't get shrunk. Those scenarios work fine now that we
do the shrinking before the rebalance). This ends up simplifying the
code, because the rebalance algorithm doesn't have to do as much work.