Add SVG of a green tree to docs
Source: https://pixabay.com/vectors/nature-plant-tree-2027899/
License: Pixabay license, text below retrieved 2020-01-28

Images and Videos on Pixabay are made available under the Pixabay License on the following terms. Under the Pixabay License you are granted an irrevocable, worldwide, non-exclusive and royalty free right to use, download, copy, modify or adapt the Images and Videos for commercial or non-commercial purposes. Attribution of the photographer or Pixabay is not required but is always appreciated.

The Pixabay License does not allow:

1.  sale or distribution of Images or Videos as digital stock photos or as digital wallpapers;
2.  sale or distribution of Images or Videos e.g. as a posters, digital prints or physical products, without adding any additional elements or otherwise adding value;
3.  depiction of identifiable persons in an offensive, pornographic, obscene, immoral, defamatory or libelous way; or
4.  any suggestion that there is an endorsement of products and services by depicted persons, brands, and organisations, unless permission was granted.

Please be aware that while all Images and Videos on Pixabay are free to use for commercial and non-commercial purposes, depicted items in the Images or Videos, such as identifiable people, logos, brands, etc. may be subject to additional copyrights, property rights, privacy rights, trademarks etc. and may require the consent of a third party or the license of these rights - particularly for commercial applications. Pixabay does not represent or warrant that such consents or licenses have been obtained, and expressly disclaims any liability in this respect.