Fix failing test by handling owner token correctly
Owner tokens should always be compared with object identity, not with
value equality, and the failing test found the one place in the code
where I had compared by value equality. This used to pass when all
generated tokens pointed to different strings, but now that they all
point to the same string (for efficiency), it fails. Comparing by object
identity, as we're supposed to, works.