Update to new scaffolding version
Update to latest MiniScaffold version (0.34) to bring in improved build
(a project instead of a build.fsx script, which means easier tooling and
more flexibility in using NuGet packages). This change also changes the
way docs are built from having a docsTool inside the project to using
the fsdocs tool.