Fix error FS0410 about type accessibility
When testing an internal class like RRBTransientVector, the code
accessing it also needs to be marked internal, otherwise the F# compiler
complains about a public method accessing an internal method/class. In
this case it's only being accessed from a unit test and won't ever be
exposed to the outside world, but the compiler can't know that. Mark
all the relevant test code internal as well, and the compiler is happy.