Move to .NET 6 for build (#1)
* Update to new scaffolding version

Update to latest MiniScaffold version (0.34) to bring in improved build
(a project instead of a build.fsx script, which means easier tooling and
more flexibility in using NuGet packages). This change also changes the
way docs are built from having a docsTool inside the project to using
the fsdocs tool.

* Disable code format checking for now

We'll get the build working first, then fix formatting issues

* Don't generate Program.fs in test project

.NET now autogenerates Program.fs if your test project doesn't contain
one, so if you had named it Main.fs then your entrypoint won't be the
last file of the project any longer. This can be suppressed with the
GenerateProgramFile property in your .fsproj file.

* Fix typo and silence build warning

* Fix error FS0410 about type accessibility

When testing an internal class like RRBTransientVector, the code
accessing it also needs to be marked internal, otherwise the F# compiler
complains about a public method accessing an internal method/class. In
this case it's only being accessed from a unit test and won't ever be
exposed to the outside world, but the compiler can't know that. Mark
all the relevant test code internal as well, and the compiler is happy.

* Stress-test script now specifies net6.0

* Let tests see library internals

Since the tests need to access RRBTransientVector in order to test its
thread safety, we need the InternalsVisibleTo attribute.